requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
chardet>=5.0.0
